Pass MergedStores to endpoint/jobs rather than only staging store
This patch changes the code so that the MergedStores object is known in the
endpoints and job execution code.

This is necessary, because the jobs must be able to load artifacts from the
release store as well, for example if a library was released in another submit
and we can reuse it because the script for that library didn't change.

For that, the interface of StoreRoot::join() was changed

    - -> Result<FullArtifactPath<'a>>
    + -> Result<Option<FullArtifactPath<'a>>>

whereas it returns an Ok(None) if the artifact requested does not exist.
This was necessary to try-joining a path on the store root of the staging store,
and if there is no such file continue with try-joining the path on the release
store.
The calling code got a bit more complex with that change, though.

Next, the MergedStores got a `derive(Clone)` because clone()ing it is cheap
(two `Arc`s) and necessary to pass it easily to the jobs.

Each instance of the code where the staging store was consulted was changed to
consult the the release store as well.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>