Implement patch-path rewriting
This patch ensures that pathes for patches are always relative to the repository
root, when loading the packages layered from the repository.

The problem is well described in the comment in the code, so this commit message
ends here.

This patch was written using an proposed feature for the config-rs dependency:
https://github.com/mehcode/config-rs/pull/165

This PR added a `Config::get_value_mut()` function and exported the
`config::value::ValueKind` type in config-rs public API.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>