Multiple release stores
This patch adds the ability to have more than one release store.

With this patch, a user can (has to) configure release store names in the
configuration file, and can then specify one of the configured names to release
the artifacts to.

This way, different release "channels" can be served, for example a stable
channel and a rolling release channel (although "channel" is not in our wording).

The code was adapted to be able to fetch releases from multiple release
directories, in the crate::db::find_artifact implementation, so that re-using
artifacts works across all release directories.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>