Remove relative speed setting, select endpoint by utilization instead
This patch removes the "speed" setting from the configuration, which was
introduced to set a relative speed for each endpoint, with the idea that the
scheduler then would select a faster node preferably.

Instead, the utilization of an endpoint is now calculated (number of running
jobs vs allowed maximum jobs on the endpoint), and the endpoint with lower
utilization is selected.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>