Add subcommand "endpoint <ENDPOINT> ping"
This is the first "endpoint" subcommand implementation, which adds a "ping"
subcommand, that can be used to ping all endpoints.

This is the first step towards a CLI interface for endpoint maintenance
commands. "ping" is the least complex one to implement and the docker crate
("shiplift") offers that functionality where one can ping the docker endpoints
for availability, thus this is chosen as first implementation.

Later, container cleanup and so on will be added as well, based on this work.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>