Re-write patch postprocessing
This patch rewrites patch post-processing with a new setter in the config-rs
crate:

    Config::set_once()

That is introduced on a PR I've written for the crate.
The PR: https://github.com/mehcode/config-rs/pull/172

The new function is more likely to land on the config-rs master branch than the
old one, as it does not introduce more exported types from the config-rs crate.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>