Fix: Error out if a patch file is missing
This patch fixes a bug where a patch file is not there.
Before this patch, we were simply ignoring non-existing files in the iterator,
because during development of this algorithm, it seemed to be the right idea
because of the recursion that is happending.

The patch-branch-patching that is happening in the recursion, that rewrites the
pathes to the patches during the recursive loading of the packages, used to
yield invalid pathes at some point, which simply could be ignored. That happened
before that patch.

But because during the development of the recursive loading, the scheme how this
all works was changed, it does not yield invalid pathes anymore.
Hence, we can be sure that either the file is here or it is not - which is an
error then.

I have to say that I'm not particularly good with recursion, but as far as my
tests go, this seems to work as intended now.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>