Rewrite jobs-env-filtering
This patch rewrites the filtering for environment variables in the "db jobs"
subcommand.

The output does not contain the environment variables anymore, which is a
drawback, but the implementation is _way_ more easy to understand, which I
consider an bigger advantage.

The implementation now actually queries the database twice: first it gets all
job ids from the JobEnv->EnvVar mapping and then builds a filter based on these
IDs for the job output.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>