Shuffle endpoint configurations
As the comment in the code explains, we're always loading the endpoints from the
same file and feed it to the scheduler the same way.
Because of that, we always schedule to the first endpoint until it is full and
then to the second.

To have a bit more spreading over the endpoints, shuffle the configurations
before connecting, so this way the scheduler does not always (each time butido
is called) use the same endpoint as first endpoint.

This is not a perfect solution, but a simple and working one.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>