Filter out the /outputs/ directory
The issue here is that we copy all build results (packages) in the container to
/outputs and then butido uses that directory to fetch the outputs of the build.

But, because how the docker API works, we get a TAR stream from docker that
_contains_ the /outputs directory. But of course, we don't want that.

Until now, that was no issue. But it has become one now that we start adopting
butido for our real-world scenarios.

This patch adds filtering out that /outputs portion of the pathes from the tar
archive when writing all the things to disc.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>