Refactor: Database connection setup code
This patch refactors the database connection setup code to involve less copying
of data in memory.

It also changes the configuration value type of the database port and the
database connection timeout setting to be `u16` instead of `String`, which is
another improvement.

Because the `DbConnectionConfig` type no longer holds owned values, it must be
annotated with a lifetime. Usages were adapted.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>