Add endpoint name and container id in progress message
This patch adds the endpoint name and a (7 character) appreviated container id
in the progress message, so that the user can see where the package is built.

This reduces the compile+debug loop by one step in the case where the user does
not want to inspect the logs anyways, because they can jump to the right
container right away and do not have to `butido db job -sL <job id>` to find out
where the job was running.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>