Fix: git commit hash/author information passing
This fixes the git commit hash and author information passing.
The changes in 2d72cbed2495517dba84ec4d46e5f521ff46412b did not add the
environment variables to the `find_artifact()` call when searching for
replacement artifacts.

Thus, calling the same build twice, with the staging directory from the first
call added in the second call, resulted in a full rebuild, although the
artifacts from the first call could be resused.

With this patch applied, this issue is fixed by adding the environment variables
to the environment passed to the `find_artifacts()` call.

Fixes: 2d72cbed2495517dba84ec4d46e5f521ff46412b ("Add feature to pass git author and git commit information to container")
Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>