Remove sourcehut builds
Since today, sourcehut builds should only work if payment is enabled.
Right now it still seems to work. I am not sure whether this is a bug or
intended. Either way, to be nice, remove the build setup for now until
we figured out what is happening and how to continue.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>