Try to parse time with " 00:00:00" appended
This makes it a bit easier to only specify a date.
If the parsing fails (e.g. with "2020-01-01"), this patch adds another step of
parsing by adding " 00:00:00" to the commandline argument and trying again to
parse.

This way, the CLI argument can be "2020-01-01" and the time does not have to be
specified.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>