Revert "Add support for filtering releases by package name and version"
Filtering the releases from the database should not be done via a regex in the
application code. Rationale is that filtering with a regex is only wanted if the
user _searches_ for something. In this instance, the user is expected to know
already what they want to see, so this is not a _search_, but a _filtering_.

Hence, we rather implement the filtering via an appropriate database query
instead of loading all releases and then filtering by regex.

This reverts commit b665c7cd4aa4ea900dc194a6ba5ad9de527f978c.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>