Revert "Fix: Make patch path absolute"
This fix was actually buggy.

The problem here is that the `Package` struct is used to fill variables in the
script. If we overwrite the patch pathes with their absolute variant, the script
gets interpolated with the absolute paths, resulting in things like

    patch -p0 < /patches//home/user/package_repository/software/p.patch

but the copy-to-container process copies (should copy) the patch file to

    /patches/software/p.patch

This reverts commit dcc8b42810d34b2b906f0307c26c4826d16ad997.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>