Fix package print format default
Maybe this is necessary because of the handlebars update, I am not entirely sure
(because it is actually a not-so-much-used feature). This patch styles the
default format string for the output to be more visually pleasing.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>
(cherry picked from commit a1e759348fe9184ec6f234a91d1d6d80468cdcfa)