Rewrite BuildDependency type as enum
This patch reimplements the BuildDependency type as an enum.

This is required because later we want to be able to use an "if" expression in
the build dependencies, for example (not decided yet) something like

    build = [
        {
            d = "libfoobar",
            if = [ { has_env = [ "TARGET_RH7" ]; } ]
        }
    ]

(no valid toml here, formatting for readability)

to make a dependency optional on a condition.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>