Add CLI parameters for tree-of subcommand to add condition-checking data
This patch adds parameters for the tree-of subcommand which can be used to
change the package-DAG-building based on conditional requirements.

If a package has a dependency-condition that only includes a dependency if there
is a certain image used to build, the "image" parameter can be used to inspect
the package tree (DAG) build with that image in mind.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>