Fix: Propagate information whether an Artifact was built or reused
This patch fixes a bug where butido did not rebuild a package if one of its
dependencies was rebuild.

Of course, if in a dependency chain with libA -> libB, if libB gets rebuild, we
need to rebuild libA as well.
For this, a new (Wrapper-)type `ProducedArtifact` was added (private to the
orchestrator module) that holds that information.

This information is only necessary between the individual build workers. If we
know from all dependencies that the artifacts were reused, we can check for a
similar job in the database and reuse artifacts from that job as well. If one
dependency was built, we need to rebuild the current package as well.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
(cherry picked from commit 507aeeb8899a714fe794814615b61eeef91c7ed2)