Pin transitive dependency "zeroize"
dialoguer depends on zeroize, which broke their public interface in 1.4.0 by
updating the minimum required rust version to 1.51.0.
Because we have 1.50.0, this does not work anymore.

Because dialoguer depends on zeroize "1.1.1", we can pin zeroize to 1.3 in our
dependencies, to force the build to not use the 1.4.* release (or newer) of
zeroize.

This is of course not a long-term solution. The proper fix is to update our
MSRV, but I do not want to do that right before our 0.2.0 release.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>