Schedule to host where fewer containers are running
This is not perfect, of course, as we rather should check the load on the host
rather than some artificial number of running containers (more failed jobs
result in more running containers, which results in more scheduled to the host),
but still, this is an improvement over the previous state because utilization
did not express enough.

So the issue was that the utilization only represents the utilization of an
endpoint _from the current process_.

If butido is called three times, the same host was selected for a job because
the utilization in each butido process resulted in the same values for this
endpoint.

In practice, this resulted in a build for "gcc" to be scheduled to the same
host when butido was executed three times for "gcc" (e.g. for different
platforms).

The endresult was a load of ~70 on a host where 24 was "100% load".

This is not ideal. Checking the load of a host is not possible via the docker
API, so this is what I came up with. Now, builds are scheduled to endpoints
which have fewer running containers.

This might not result in a perfect distribution (or even a "good" one), but
still it might result in a better distribution of workload than in before this
patch.

Some completely non-artificial testing resulted in the expected behaviour.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Tested-by: Matthias Beyer <matthias.beyer@atos.net>