Fix clippy: Use BTreeMap instead of HashMap
clippy complained that we cannot have a derived PartialEq and a handwritten Hash
implementation.

HashMap does not implement Hash.  After some research we found that BTreeMap
implements Hash, so we can derive Hash if we use BTreeMap as a key-value type
for the `env_eq` member.
Logically they are the same and this reduces code size, so go for it.

This has the nice benefit, that we do not need to implement PartialOrd and Ord
either.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>