Remove myself
This patch removes me (with the atos.net mail address) from the codebase, as I
quit my job at 31.12.2021.

Thus, this patch changes the human_panic setup call to refer to
science-computing ag, opensoftware, which is the team that I belong(ed) to.

My private mail address remains in the codebase (in the authors), as I still
authored the project. The atos.net mail address is removed because I am no
longer reachable under this address, though.

Christoph comes first in the authors field now, though.

Signed-off-by: Matthias Beyer <matthias.beyer@atos.net>
Signed-off-by: Matthias Beyer <mail@beyermatthias.de>
CC: Christoph Prokop <christoph.prokop@atos.net>