Update dependency: clap: 3.0.0-beta.2 -> ^3.0.0
This patch updates the clap dependency and the clap_generate dependency
(which was renamed to clap_complete) from the beta versions to 3.0.0.

The patch is a bit big, but that's only due to the fact that some things
were renamed in clap.
The only change that is not a simple rename is the clap_complete
interface changes, which should still be functionally equivalent though.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>