Rewrite value fetching
Between clap 3.0 and clap 4.0, the way how values are fetched from the
CLI has changed. Notably the Arg::value_of and Arg::values_of functions
were replaced by something that can actually be a bit more typed.

This patch rewrites all these uses and introduces a bit more typing in
the CLI definition itself.

Types like integers or uuid::Uuid are moved to the CLI definition,
because these are trivial to make into clap ValueParser types.

More types should and will be transformed to be able to catch errors
already in CLI parsing instead of down the road during the execution of
butido.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>