Fix for indicatif interface changes
So we need to adapt for the indicatif interface changes.
First of all, `ProgressStyle::template()` returns a `Result` now, so we
need to adapt `ProgressBars::bar()` to also return a `Result` type and
adjust all uses of this function.

Secondly, we do not need to `MultiBar::join()` anymore, which makes some
future collecting less complex.

Signed-off-by: Matthias Beyer <mail@beyermatthias.de>