Include the package name and version in the logfile name
This makes it easier / more convenient for package developers to find
the right log file as it avoids the need to lookup the UUID first.

We decided to prepend the package name and version so that the shell's
tab completion functionality can be used to discover the filenames and
globbing could also be useful.

This implements most of #50 (we just don't include the short Linux
distribution name yet).

Signed-off-by: Michael Weiss <michael.weiss@atos.net>
Co-authored-by: Nico Steinle <nico.steinle@atos.net>
Tested-by: Michael Weiss <michael.weiss@atos.net>