Shorten the header fields of the db jobs output
We decided to shorten the header fields by default so that we can add
more columns in the future without exceeding the available line width.
Fixes #49.

Note: We only use the shortened header field names in the DB jobs
output (table) so that the rest of the code still uses the same field
names as in the structs and DB model.

Signed-off-by: Nico Steinle <nico.steinle@atos.net>
Co-authored-by: Michael Weiss <michael.weiss@atos.net>