actions: Run optional clippy (linting) checks with the beta toolchain
More recent toolchain versions come with additional clippy lints.
E.g., the current stable version is 1.65.0 and shows three additional
clippy warnings (compared to 1.64.0 / the MSRV). We already saw those
warnings locally (until aa10870).

We only make the check with the MSRV required, as contributors shouldn't
have to fix clippy warnings in other parts of the code (unrelated to
their PR; we'll fix any additional warnings when bumping the MSRV).

However, it is useful to also run the clippy check with a more recent
toolchain version so that we maintainers can noticed and fix new
warnings in advance (without having to run clippy locally).
This is implemented so that one sees the failing optional job (hopefully
that won't confuse contributors) but the "CI" job that is required for
bors will still pass.
Most contributors will likely use the stable version so using the beta
version for the check seems ideal to catch warnings in advance (so that
contributors don't see them when they run clippy locally).

Relevant GitHub actions documentation:
https://docs.github.com/en/actions/using-workflows/workflow-syntax-for-github-actions#jobsjob_idcontinue-on-error

Signed-off-by: Michael Weiss <michael.weiss@atos.net>