Enable caching across jobs of the CI workflow
The `Swatinem/rust-cache@v2` action uses a prefix as part of the cache
key. By default, this prefix is formed by combining the `prefix-key`
(defaults to "v0-rust") and the `shared-key` (defaults to the job name).
As a result, this breaks caching within the CI workflow as the prefix
differs for each job (e.g., `v0-rust-test` for the test job) which
prevents the cache from being restored (`No cache found.`).

Setting `shared-key` (default: empty) will prevent this by using that
string instead of the job name.
Therefore, caching within a ci.yml workflow will now work as expected.

Signed-off-by: Michael Weiss <michael.weiss@atos.net>