Cleanup and unify the cargo CLI options
The `--all-targets` option is equivalent to specifying `--lib --bins
--tests --benches --examples`. We seem to only need `--tests` (as the
binaries are built as unittests by default) but we might as well set
`--all-targets` for future-proofing (in case we ever add any examples,
benchmarks, or a library).
By running `cargo check` with `--all-targets` we ensure that all of the
code is already compiled for subsequent commands (e.g., useful for the
caching of the CI workflow).

The `--all` option is a deprecated alias for `--workspace`. We only have
one member (butido) in the workspace or rather don't define a workspace
in Cargo.toml (it would have a `workspace` table instead of the
`package` table). Therefore, we may drop that option for simplicity.

The same applies to the `--all-features` option: We don't define any
butido features in Cargo.toml (no `features` table/section).

Last but not least, I updated CONTRIBUTING.md to (almost) match our CI
workflow.

Signed-off-by: Michael Weiss <michael.weiss@atos.net>