Run the cargo-deny CI job independent of the check job
We probably don't benefit of the caching here. The cargo-deny job
doesn't use `dtolnay/rust-toolchain` as
`EmbarkStudios/cargo-deny-action` uses a Docker container (currently
based on `rust:1.65.0-alpine3.16`). This can/will result in a different
cache key (currently due to a different compiler version and the lack of
the `CARGO_TERM_COLOR` environment variable).
Plus cargo-deny doesn't need to compile any packages. It needs to
download a few things (in /home/runner/.cargo - which doesn't seem to
get bind mounted into the Docker container anyway) but
`swatinem/rust-cache` also needs to download the cache.

This change should make things faster as the cargo-deny job doesn't need
to wait for the check job anymore and the caching likely doesn't work at
all (it likely won't matter though as the cargo-deny job shouldn't be
the slowest job of the CI workflow but let's drop the caching here for
simplicity and less confusion).

Signed-off-by: Michael Weiss <michael.weiss@atos.net>