Update all dependencies
This is just the result of running `cargo update` without touching
Cargo.toml. It might interfere with dependabot but I think it's a good
idea as this hasn't been done for quite a while and should hopefully
also make it easier to manually update some dependencies that dependabot
cannot handle due to complex interdependencies (and to resolve some
other dependency issues / make updates smoother).

Signed-off-by: Michael Weiss <michael.weiss@atos.net>