Support specifying short names as aliases for container image names
This changes the `images` configuration setting so that shorter names
can be assigned to container images. This is done in preparation of
adding an additional column to the `butido db jobs` output that shows
the image that was used for the build (#48). The full image names are
often too long so it should be possible to specify shorter names as
aliases. The `compatibility` setting is bumped since this is a breaking
change that requires updating the configuration files (side note: we
should probably use a normal number for the `compatibility` setting as
it shouldn't depend on the butido version).

Fix #101.

Signed-off-by: Michael Weiss <michael.weiss@atos.net>